# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventManagement swagger application",
        "description": "Api d'une Application de  gestion des evenements ",
        "contact": {
            "email": "taneyemc@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/guest/addGuest": {
            "post": {
                "tags": [
                    "guest"
                ],
                "summary": "Add new guest",
                "description": "Create a new guest",
                "operationId": "887636e20659c850684a20040aba837c",
                "requestBody": {
                    "description": "Guest details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "email",
                                    "phone",
                                    "place_id",
                                    "table_id",
                                    "event_id"
                                ],
                                "properties": {
                                    "firstname": {
                                        "description": "Guest's firstname",
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "description": "Guest's lastname",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Guest's email",
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "Guest's gender",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Guest's phone",
                                        "type": "string"
                                    },
                                    "place_id": {
                                        "description": "Guest's place_id",
                                        "type": "string"
                                    },
                                    "table_id": {
                                        "description": "Guest's table_id",
                                        "type": "string"
                                    },
                                    "event_id": {
                                        "description": "Guest's event_id",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guest added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Guest added successfully"
                                        },
                                        "guest": {
                                            "$ref": "#/components/schemas/Guest"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "firstname": [
                                                    "The firstname field is required."
                                                ],
                                                "email": [
                                                    "The email field is required.",
                                                    "The email field must be in the form of an email address."
                                                ],
                                                "phone": [
                                                    "The phone field is required."
                                                ],
                                                "place_id": [
                                                    "The place_id field is required."
                                                ],
                                                "table_id": [
                                                    "The table_id field is required."
                                                ],
                                                "event_id": [
                                                    "The event_id field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/guest/updateGuest/{guest}": {
            "put": {
                "tags": [
                    "guest"
                ],
                "summary": "Update an guest",
                "description": "Update an existing guest by ID",
                "operationId": "507b4d56a0f89353d8a22b77f230289f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the guest",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Guest details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "firstname": {
                                        "description": "Guest's firstname",
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "description": "Guest's lastname",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Guest's email",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Guest's phone",
                                        "type": "string"
                                    },
                                    "place_id": {
                                        "description": "Guest's place_id",
                                        "type": "string"
                                    },
                                    "table_id": {
                                        "description": "Guest's table_id",
                                        "type": "string"
                                    },
                                    "event_id": {
                                        "description": "Guest's event_id",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Guest updated successfully"
                                        },
                                        "guest": {
                                            "$ref": "#/components/schemas/Guest"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Guest not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "firstname": [
                                                    "The firstname field is required."
                                                ],
                                                "email": [
                                                    "The email field is required.",
                                                    "The email field must be in the form of an email address."
                                                ],
                                                "phone": [
                                                    "The phone field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/guest/getAllGuest": {
            "get": {
                "tags": [
                    "guest"
                ],
                "summary": "Get all guest",
                "description": "Retrieve a list of all guest",
                "operationId": "33262a842b83a11bcff343654bccd52d",
                "responses": {
                    "200": {
                        "description": "List of all guest",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Guest"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No owners found"
                    }
                }
            }
        },
        "/api/guest/showGuest/{guest}": {
            "get": {
                "tags": [
                    "guest"
                ],
                "summary": "Show guest detail by ID",
                "description": "Retrieve details of a specific guest by ID",
                "operationId": "8abd5a8dec8807e4e8d2c0194f189016",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the guest",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guest details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Guest": {
                "required": [
                    "firstname",
                    "email",
                    "phone",
                    "place_id",
                    "table_id",
                    "event_id"
                ],
                "properties": {
                    "firstname": {
                        "description": "Guest's firstname",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "Guest's lastname",
                        "type": "string"
                    },
                    "email": {
                        "description": "Guest's email",
                        "type": "string"
                    },
                    "gender": {
                        "description": "Guest's gender",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Guest's phone",
                        "type": "string"
                    },
                    "status": {
                        "description": "Guest's status",
                        "type": "enum"
                    },
                    "place_id": {
                        "description": "Guest's place_id",
                        "type": "string"
                    },
                    "table_id": {
                        "description": "Guest's table_id",
                        "type": "string"
                    },
                    "event_id": {
                        "description": "Guest's event_id",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}